Eliminar suscripción, alertas

+++ Frontend/src/app/navigation-products/products/products.component.ts
@@ -3,7 +3,7 @@ import { EquipoService, filter, traerProducto, subscribe, requestSubscriptions, 
 import { PageEvent } from '@angular/material/paginator';
 import { FormControl, FormGroup, Validators} from '@angular/forms';
 import { Router} from '@angular/router';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-products',
@@ -81,6 +81,7 @@ export class ProductsComponent implements OnInit {
         var request:requestSubscriptions = <any>res
         this.UserSubscription = request.msg
 
+        this.UserSubscriptionID = []
         for(var index in this.UserSubscription){
           this.UserSubscriptionID.push(this.UserSubscription[index].id_product_category)
         }
@@ -149,8 +150,17 @@ export class ProductsComponent implements OnInit {
 
     if(this.subscribed === false){ /// Si no esta suscrito
       this.equipoService.addsubscription(this.nuevaSuscripcion).subscribe(res=>{
-        console.log(this.nuevaSuscripcion)
         var info:BookInfo = <any>res
+
+        /// Alerta
+        Swal.fire({
+          position: 'center',
+          icon: 'success',
+          title: info.msg,
+          showConfirmButton: false,
+          timer: 1500
+        })
+
         console.log(info.msg)
 
         this.subscribed = true
@@ -160,15 +170,41 @@ export class ProductsComponent implements OnInit {
       })
 
     }else{ /// Si esta suscrito, anular suscripción
-      this.equipoService.deleteSubscription(this.nuevaSuscripcion).subscribe(res=>{
-        var info:BookInfo = <any>res
-        console.log(info.msg)
 
-        this.subscribed = false
-        this.updateSubscriptionList()
-      }, error =>{
-        console.log(error)
+      /// Alerta cancelar sub
+      Swal.fire({
+        title: '¿Anular la suscripción de ' + this.UserSubscription[this.UserSubscriptionID.indexOf(+this.nuevaSuscripcion.fk_id_product_category)].var_name + '?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        cancelButtonText: 'No',
+        confirmButtonText: 'Anular'
+      }).then((result) => {
+        if (result.isConfirmed) {
+
+          /// Llama a la funcion de borrar suscripción
+          this.equipoService.deleteSubscription(this.nuevaSuscripcion).subscribe(res=>{
+            var info:BookInfo = <any>res
+            console.log(info.msg)
+    
+            this.subscribed = false
+            this.updateSubscriptionList()
+          }, error =>{
+            console.log(error)
+          })
+
+          Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: "Suscripción anulada",
+            showConfirmButton: false,
+            timer: 1500
+          })
+        }
       })
+
+
     }
     
   }
